# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: SIRCCD API
  description: |
    API REST para el Sistema Inteligente Urbano para Reporte y Priorización de Daños Viales.
    
    Permite gestionar reportes ciudadanos, clasificación automática mediante IA, 
    deduplicación, priorización multicriterio y gestión operativa municipal.
  version: 1.0.0
  contact:
    name: Equipo SIRCCD
    email: soporte@sirccd.gob.sv
servers:
  - url: http://localhost:8000/api/v1
    description: Desarrollo local
  - url: https://api.sirccd.gob.sv/api/v1
    description: Producción
...
paths:
...
  /reportes:
    get:
      tags:
        - reportes
      summary: Listar reportes
      description: Obtiene lista de reportes con filtros opcionales
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: estado
          in: query
          schema:
            type: string
            enum: [pendiente, procesando, validado, rechazado, duplicado]
        - name: tipo_dano
          in: query
          schema:
            type: string
            enum: [bache, grieta, hundimiento]
        - name: fecha_desde
          in: query
          schema:
            type: string
            format: date
        - name: fecha_hasta
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Lista de reportes
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                    example: 150
                  page:
                    type: integer
                    example: 1
                  limit:
                    type: integer
                    example: 20
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Reporte'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - reportes
      summary: Crear nuevo reporte
      description: Crea un reporte ciudadano con foto y geolocalización
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - imagen
                - latitud
                - longitud
              properties:
                imagen:
                  type: string
                  format: binary
                  description: Foto del daño vial
                latitud:
                  type: number
                  format: double
                  example: 13.692940
                longitud:
                  type: number
                  format: double
                  example: -89.218191
                descripcion:
                  type: string
                  example: Bache grande en entrada de colonia
                direccion:
                  type: string
                  example: Calle Principal, Col. Escalón
      responses:
        '201':
          description: Reporte creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reporte'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /reportes/{id}:
    get:
      tags:
        - reportes
      summary: Obtener reporte por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Detalles del reporte
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reporte'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    patch:
      tags:
        - reportes
      summary: Actualizar estado del reporte
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                estado:
                  type: string
                  enum: [pendiente, procesando, validado, rechazado, duplicado]
                observaciones:
                  type: string
      responses:
        '200':
          description: Reporte actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reporte'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ==========================================
  # INCIDENTES - Gestión de incidentes validados
  # ==========================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Reporte:
      type: object
      properties:
        id:
          type: string
          format: uuid
        usuario_id:
          type: string
          format: uuid
        imagen_url:
          type: string
          format: uri
        imagen_procesada_url:
          type: string
          format: uri
        latitud:
          type: number
          format: double
        longitud:
          type: number
          format: double
        direccion:
          type: string
        descripcion:
          type: string
        tipo_dano:
          type: string
          enum: [bache, grieta, hundimiento]
          nullable: true
        confianza_clasificacion:
          type: number
          format: float
          nullable: true
        severidad:
          type: number
          format: float
          nullable: true
        estado:
          type: string
          enum: [pendiente, procesando, validado, rechazado, duplicado]
        es_duplicado:
          type: boolean
        duplicado_de:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
  responses:
    BadRequest:
      description: Solicitud inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: bad_request
            message: La solicitud no cumple con los requisitos
    Unauthorized:
      description: No autorizado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: unauthorized
            message: Token de autenticación inválido o expirado
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: not_found
            message: El recurso solicitado no existe
    ValidationError:
      description: Error de validación
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  details:
                    type: array
                    items:
                      type: object
                      properties:
                        field:
                          type: string
                        message:
                          type: string
          example:
            error: validation_error
            message: Los datos proporcionados no son válidos
            details:
              - field: email
                message: Formato de email inválido
              - field: password
                message: La contraseña debe tener al menos 8 caracteres
security:
  - bearerAuth: []